move tests to the correct location
